Extract service from `CompletedFileReviewJob`
To follow the pattern of a job class having a corresponding service
object.

https://trello.com/c/fLtPcESw
